Move test_not_loading_targets_twice
Move test_not_loading_targets_twice to
test_updater_top_level_update.py.
